Define network protocol
