+++ schemas/push.ts
@@ -5,19 +5,15 @@ export const mutationSchema = z.object({
   id: z.number(),
   name: z.string(),
   args: jsonSchema,
+  // TODO: Will be required soon.
+  timestamp: z.number().optional(),
 });
 
 export const pushRequestSchema = z.object({
-  id: z.string(),
   mutations: z.array(mutationSchema),
   pushVersion: z.number(),
   schemaVersion: z.string(),
 });
 
-export const pushResponseSchema = z.object({
-  id: z.string(),
-});
-
 export type Mutation = z.infer<typeof mutationSchema>;
 export type PushRequest = z.infer<typeof pushRequestSchema>;
-export type PushResponse = z.infer<typeof pushResponseSchema>;
